0.7.5 update scripts gm (#60)
