+++ scripts_gm/README.md
@@ -1,0 +1,15 @@
+# scripts_gm
+
+使用掘金量化终端的机器交易员
+
+## 缩写
+
+* rt -> realtime -> 实盘
+* bt -> backtest -> 回测
+
+## 环境配置
+
+直接从github安装最新版 `CZSC`：
+```
+pip install git+git://github.com/zengbin93/czsc.git -U
+```
