+++ scripts_gm/src/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+author: zengbin93
+email: zeng_bin8888@163.com
+create_dt: 2021/8/3 17:09
+"""
+import czsc
+
+version = "0.1.2"
+update_date = '20210828'
+
+# czsc 版本校验
+assert czsc.__version__ >= '0.7.5'
+assert czsc.__date__ >= '20210828'
