+++ scripts_gm/src/tactics/__init__.py
@@ -2,5 +2,7 @@
 """
 author: zengbin93
 email: zeng_bin8888@163.com
-create_dt: 2021/8/3 17:47
+create_dt: 2021/8/3 17:43
 """
+
+from .share import TacticShareA
